+++ tests/test_pipeline.py
@@ -49,6 +49,7 @@ def test_pipeline():
                     "table_source": {
                         "drops": None,
                         "filter": None,
+                        "broadcast": False,
                         "read_as_stream": True,
                         "renames": None,
                         "selects": None,
@@ -153,6 +154,7 @@ def test_pipeline():
                     "table_source": {
                         "drops": None,
                         "filter": "created_at = '2023-09-01T00:00:00Z'",
+                        "broadcast": False,
                         "read_as_stream": True,
                         "renames": None,
                         "selects": None,
@@ -178,6 +180,7 @@ def test_pipeline():
                                         "value": {
                                             "drops": None,
                                             "filter": None,
+                                            "broadcast": False,
                                             "read_as_stream": True,
                                             "renames": {"symbol2": "symbol"},
                                             "selects": None,
@@ -226,6 +229,7 @@ def test_pipeline():
                                         "value": {
                                             "drops": None,
                                             "filter": None,
+                                            "broadcast": False,
                                             "read_as_stream": True,
                                             "renames": None,
                                             "selects": None,